python/otel-python-mcp-examples/simple-client-server/client/pyproject.toml
[project]
name = "python-mcp-client-foo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.53",
    "debugpy>=1.8.17",
    "mcp[cli]>=1.17.0",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-distro>=0.58b0",
    "opentelemetry-exporter-otlp>=1.37.0",
    "pip>=25.2",
    "requests>=2.32.5",
]
